Fix eslint warning react-refresh/only-export-components
