+++ src/components/Sorting.tsx
@@ -8,13 +8,7 @@ import {
   useMediaQuery,
   useTheme,
 } from '@mui/material';
-
-export enum SortingOptions {
-  TITLE = 'Order by',
-  NAME = 'Name',
-  LAST_UPDATED = 'Last Updated',
-  CREATED_DATE = 'Created',
-}
+import { SortingOptions } from '../types';
 
 interface SortingProps {
   filteredAndSortedProjectsLength: number;
